Move config file to src folder
