+++ front/src/utils/HttpClient.js
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import config from '../../config';
+import config from '../config';
 
 const MAX_RETRY = 3;
